requirements.txt
django
djangorestframework
django-cors-headers
psycopg2-binary
python-dotenv
stripe
dj-database-url
whitenoise
gunicorn
djangorestframework-simplejwt
